autoinfra-backend/requirements.txt
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.1
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.2
